Add files via upload
This task involves conducting exploratory data analysis on a considerable dataset from an online retail store, with the objective of simulating the work of a business manager. The primary goal is to pinpoint areas of the business that are underperforming and devise strategies to transform these challenges into opportunities for increased profitability. By investigating the data, we seek to extract valuable insights that can inform us about specific business problems and illuminate pathways to harness them as growth opportunities.